// Client/bindings/FarmFieldGroupMeta.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FarmFieldMeta } from "./FarmFieldMeta";

export type FarmFieldGroupMeta = {
  id: number;
  farm_id: number;
  name: string;
  draw_color: string;
  fields: Array<FarmFieldMeta>;
};